로또 3단계 리팩토링 (#309)
* docs : 2단계 요구사항 정리

* chore : apache common collections 의존성 추가

* refactor: [3단계] 일급 컬렉션 적용, 로또 티켓 생성 방식 변경

- LottoNumber, LottoNumbers 클래스 추가
- RandomLottoTicket 클래스 추가
- 기존 Factory 인터페이스에서 로또 티켓을 생성하는 방식에서 각 생성자에서 로또 티켓을 생성하는 방식으로 변경
- 당첨 로또 번호들과 보너스 번호를 가지고 있는 WinningConditionon 클래스 추가

* fix : [3단계] RandomLottoTicket 끝범위 포함 안되는 현상 수정

* docs : 3단계 구현 내용 수정

+++ src/main/java/lotto/view/OutputView.java
@@ -19,7 +19,7 @@ public class OutputView {
         System.err.println(e.getMessage() + "\n");
     }
 
-    public void printResult(Map<LottoRank, Integer> resultCounts, int earnRate) {
+    public void printResult(Map<LottoRank, Integer> resultCounts, long earnRate) {
         System.out.println("당첨 통계");
         System.out.println("---------");
         for (LottoRank result : LottoRank.values()) {
